Rename dialog fixed. Identation fixed. GtkExtra moved to GtxExtensions
